Removed Welcome components Removed Counter Moved Heroes content into HomeView
