--- xx-final/vue-heroes-v2.7/src/views/CounterView.vue
@@ -1,20 +1,0 @@
-
-<script setup lang="ts">
-import { useCounterStore } from '@/stores/counter';
-// import { useHeroesStore } from '@/stores/heroes';
-import { storeToRefs } from 'pinia';
-const store = useCounterStore();
-const { counter, doubleCount }  = storeToRefs(store);
-const { increment }  = store;
-// const { counter, increment } = store;
-
-</script>
-
-<template>
-    <div>
-        <h3>Counter: {{counter}}</h3>
-        <h3>Double count: {{ doubleCount }}</h3>
-        <button v-on:click="increment">Increment</button>
-        
-    </div>
-</template>
